feat: improve sql audit store test with singleton containers
